Refactor
- App restructure, use locations as entry point
