+++ src/javascripts/locations/ticket_sidebar.js
@@ -1,57 +1,8 @@
-/**
- *  Example ticket sidebar app
- **/
+import App from '../modules/app'
 
-import I18n from '../../javascripts/lib/i18n'
-import { resizeContainer, render } from '../../javascripts/lib/helpers'
-import getDefaultTemplate from '../../templates/default'
+/* global ZAFClient */
+var client = ZAFClient.init()
 
-const MAX_HEIGHT = 1000
-const API_ENDPOINTS = {
-  organizations: '/api/v2/organizations.json'
-}
-
-class TicketSidebar {
-  constructor (client, appData, config) {
-    this._client = client
-    this._appData = appData
-    this._config = config
-
-    this._states = {}
-    // this._initializePromise is used to indicate app initilization(including all async operations) is complete,
-    // only used in integration testing
-    this._initializePromise = this._init()
-  }
-
-  /**
-   * Initialize module, render main template
-   */
-  async _init () {
-    const currentUser = (await this._client.get('currentUser')).currentUser
-    this._states.currentUserName = currentUser.name
-
-    I18n.loadTranslations(currentUser.locale)
-
-    const organizations = await this._client
-      .request(API_ENDPOINTS.organizations)
-      .catch(this._handleError.bind(this))
-
-    if (organizations) {
-      this._states.organizations = organizations.organizations
-
-      // render application markup
-      render('.loader', getDefaultTemplate(this._states))
-      return resizeContainer(this._client, MAX_HEIGHT)
-    }
-  }
-
-  /**
-   * Handle error
-   * @param {Object} error error object
-   */
-  _handleError (error) {
-    console.log('An error is handled here: ', error.message)
-  }
-}
-
-export default TicketSidebar
+client.on('app.registered', function (appData) {
+  return new App(client, appData)
+})
